Added the PHP LICENSE
This is the recommended license found on https://www.php.net/license/3_01.txt

I'm making the assumption that this extension, which is controlled by the PHP group, falls under this license even though it wasn't explicitly stated. I'm adding this only since PECL had a warning about one not being present when I uploaded a new version today.

If this license isn't the right one for this work, please feel free to let me know.